Replace setting PATH with a qemu_bin variable
This way we don't assume that qemu executable has to be in any
particular directory, and instead we just want a path to the binary
itself. We also provide a link to the instructions on how to build
the OVMF from source. Additionally. we remove the info function, as
it's pretty much unnecessary, as user won't be able to see it anyway.

Signed-off-by: Aodhnait Étain <aodhneine@pm.me>